feat: add UserProvider
