+++ src/Molecules/UserSelect/UserProvider/index.js
@@ -1,0 +1,3 @@
+import UserProvider from './UserProvider'
+
+export default UserProvider
